requirements.txt
pandas
seaborn
matplotlib
network
